+++ app/components/moderation-list/component.js
@@ -33,18 +33,23 @@ export default Component.extend({
         this.set('statusButtons', [
             {
                 status: 'pending',
-                iconClass: 'fa-hourglass-o icon-pending',
+                iconClass: ['fa-hourglass-o icon-pending'],
                 labelKey: 'components.moderationList.pending',
             },
             {
                 status: 'accepted',
-                iconClass: 'fa-check-circle-o icon-accepted',
+                iconClass: ['fa-check-circle-o icon-accepted'],
                 labelKey: 'components.moderationList.accepted',
             },
             {
                 status: 'rejected',
-                iconClass: 'fa-times-circle-o icon-rejected',
+                iconClass: ['fa-times-circle-o icon-rejected'],
                 labelKey: 'components.moderationList.rejected',
+            },
+            {
+                status: 'withdrawn',
+                iconClass: ['fa-circle-o icon-accepted', 'fa-minus icon-withdrawn'],
+                labelKey: 'components.moderationList.withdrawn',
             },
         ]);