+++ app/locales/en/translations.js
@@ -203,6 +203,7 @@ export default {
             pending: 'Pending',
             accepted: 'Accepted',
             rejected: 'Rejected',
+            withdrawn: 'Withdrawn',
             sort: 'Sort',
             noSubmissions: 'No submissions.',
         },
@@ -216,6 +217,8 @@ export default {
                 acceptedAutomatically: 'Accepted automatically {{timeDate}}',
                 rejectedOn: 'Rejected on {{timeDate}} by {{moderatorName}}',
                 rejected: 'Rejected {{timeDate}} by {{moderatorName}}',
+                withdrawnOn: 'Withdrawn on {{timeDate}} by {{moderatorName}}',
+                withdrawn: 'Withdrawn {{timeDate}} by {{moderatorName}}',
             },
         },
         notificationList: {
@@ -242,6 +245,7 @@ export default {
                 pending: 'submitted this {{documentType.singular}} on',
                 accepted: 'accepted this {{documentType.singular}} on',
                 rejected: 'rejected this {{documentType.singular}} on',
+                withdrawn: 'withdrew this {{documentType.singular}} on',
                 automatic: {
                     pending: 'This {{documentType.singular}} was submitted on',
                     accepted: 'This {{documentType.singular}} was automatically accepted on',
@@ -252,17 +256,22 @@ export default {
                 pendingPost: 'publicly available and searchable but is subject to removal by a moderator',
                 accepted: 'publicly available and searchable',
                 rejected: 'not publicly available or searchable',
+                withdrawn: 'not publicly available',
             },
             pending: 'pending',
             accepted: 'accepted',
             rejected: 'rejected',
+            withdrawn: 'withdrawn',
             loading: 'Loading...',
+            noReasonProvided: 'No reason provided',
             decision: {
                 makeDecision: 'Make decision',
                 modifyDecision: 'Modify decision',
+                withdrawalReason: 'View reason',
                 header: {
                     submitDecision: 'Submit your decision',
                     modifyDecision: 'Modify your decision',
+                    withdrawalReason: 'Reason for withdrawal',
                 },
                 moderator: 'Moderator',
                 base: 'This {{documentType.singular}} is',
@@ -282,6 +291,10 @@ export default {
                     label: 'Reject',
                     pre: 'Submission will not appear in search results and will remain private.',
                     post: 'Submission will be removed from search results and made private.',
+                },
+                withdrawn: {
+                    label: 'Withdraw',
+                    post: 'Submission will no longer be publicly available.',
                 },
             },
             settings: {